Extracted out common code to a separate module so it can be accessed by the macros as well
